requirements.txt
Django>=2.2.13,<3.0
djangorestframework
